Creacion de migracion, crud y modelos de permissions y permissionsRoles (#2)

+++ app/Controllers/Http/RolesController.ts
@@ -1,14 +1,16 @@
 import type { HttpContextContract } from '@ioc:Adonis/Core/HttpContext'
 import Role from 'App/Models/Role'
+import User from 'App/Models/User'
 
 export default class RolesController {
   public async index(ctx: HttpContextContract) {
-    return Role.all()
+    let roles: Role[] = await Role.query().preload('permissions')
+    return roles
   }
 
   public async store({ request }: HttpContextContract) {
     const body = request.body()
-    const newRole = await Role.create(body)
+    const newRole: Role = await Role.create(body)
     return newRole
   }
 
@@ -18,13 +20,21 @@ export default class RolesController {
 
   public async update({ params, request }: HttpContextContract) {
     const body = request.body()
-    const theRole = await Role.findOrFail(params.id)
+    const theRole: Role = await Role.findOrFail(params.id)
     theRole.name = body.name
     return theRole.save()
   }
 
   public async destroy({ params }: HttpContextContract) {
-    const theRole = await Role.findOrFail(params.id)
-    return theRole.delete()
+    let users = await User.query().where('idRole', params.id)
+    if (users) {
+      return {
+        error: 'El rol tiene usuarios asociados',
+        users: users,
+      }
+    } else {
+      const theRole: Role = await Role.findOrFail(params.id)
+      return theRole.delete()
+    }
   }
 }
